[fix] M40 GPUs can't use bfloat switch to float16 for torch_dtype

+++ src/scripts/train_lora.py
@@ -55,7 +55,7 @@ def main():
     # 🔧 use fp16 + gradient checkpointing for GPU efficiency
     model = AutoModelForCausalLM.from_pretrained(
         args.model_id,
-        dtype=torch.bfloat16,
+        torch_dtype=torch.float16,
         trust_remote_code=True,
         use_safetensors=True,
     )
